pyproject.toml
[project]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = [
    { name = "Igor Bondarenko", email = "bondarenko200021@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.0,<0.35.0",
    "pyjwt>=2.10.1,<3.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "python-multipart>=0.0.20,<0.0.21",
    "cryptography>=44.0.2,<45.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt (==4.0.1)",
    "pytest (>=8.3.5,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

